[mesmice] Jogo da Mesmice pre-release

+++ functions/src/utils/firebase.ts
@@ -3,6 +3,7 @@ import * as functions from 'firebase-functions';
 import { getFirestore, FieldValue } from 'firebase-admin/firestore';
 // Utils
 import utils from '../utils';
+import { isEmpty } from 'lodash';
 
 export const config = functions.config;
 
@@ -319,14 +320,17 @@ export const saveGame = async (
   saveContent: SaveGamePayload
 ) => {
   try {
-    if (saveContent?.set?.state) {
+    if (saveContent?.set?.state && !isEmpty(saveContent?.set?.state)) {
       await sessionRef.doc('state').set({ ...saveContent.set.state, updatedAt: Date.now() } ?? {});
     }
   } catch (error) {
     throwException(error, 'set game state');
   }
   try {
-    if (saveContent?.update?.store || saveContent?.update?.storeCleanup) {
+    if (
+      (saveContent?.update?.store && !isEmpty(saveContent?.update?.store)) ||
+      saveContent?.update?.storeCleanup
+    ) {
       const cleanup = (saveContent?.update?.storeCleanup ?? []).reduce((acc, key) => {
         if (key) {
           acc[key] = deleteValue();
@@ -340,7 +344,10 @@ export const saveGame = async (
   }
 
   try {
-    if (saveContent?.update?.state || saveContent?.update?.stateCleanup) {
+    if (
+      (saveContent?.update?.state && !isEmpty(saveContent?.update?.state)) ||
+      saveContent?.update?.stateCleanup
+    ) {
       const cleanup = (saveContent?.update?.stateCleanup ?? []).reduce((acc, key) => {
         if (key) {
           acc[key] = deleteValue();
